Add utility script to generate .env file
The rate-repository-api is running on WSL2 and therefore
Expo needs tunneling via NGROK to access it.
The NGROK isntance running on native Windows generates an
ephemeral url that is valid only for two hours or less if the
service is restarted.

To help with the tunneling configuration, this script makes
a HTTP GET request via axios to NGROK local server. If the
server is found one can read the tunnel url from response.

Finally the url is written to a .env file which can be read
with 'dotenv' and app.config.js. If .env file already exists,
only the NGROK_URL field is changed.