refetchQueries to CREATE_REVIEW mutation
If user without any reviews creates a review
it does not become visible before reloading app.
The refetch needs to be done with the variables
as Apollo Client claims to do the query with
the last variables given, which do not guarantee
the refetching of users own reviews as the default
value of 'includeReviews' is false.